requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pillow==9.0.1
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
pytz==2021.3
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.11
